+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## [Unreleased]
 
+## [1.0.2] - 2025-07-18
+    - Removed unnecessary dependencies
+    - Now watches creation and deletion of .env files within the project workspace
+    - You can now create .env files from .env.example templates
+    - Ui improvements
+    - Multiple improvements and Bug Fixes
+
 ## [1.0.1] - 2025-07-17
     - Support for adding new environment variables from the Panel
     - Support for deleting environment variables from the Panel